backend/requirements.txt
Django==3.1.7
django-cors-headers==3.7.0
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
ipdb==0.13.4
